Split out highlighting logic into file
